+++ DofusProtocol/Messages/Messages/game/inventory/exchanges/ExchangeBidHouseListMessage.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 10/28/2014 16:36:55
+// Generated on 12/29/2014 21:13:26
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -18,32 +18,32 @@ namespace Stump.DofusProtocol.Messages
             get { return Id; }
         }
         
-        public int id;
+        public short id;
         
         public ExchangeBidHouseListMessage()
         {
         }
         
-        public ExchangeBidHouseListMessage(int id)
+        public ExchangeBidHouseListMessage(short id)
         {
             this.id = id;
         }
         
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(id);
+            writer.WriteShort(id);
         }
         
         public override void Deserialize(IDataReader reader)
         {
-            id = reader.ReadInt();
+            id = reader.ReadShort();
             if (id < 0)
                 throw new Exception("Forbidden value on id = " + id + ", it doesn't respect the following condition : id < 0");
         }
         
         public override int GetSerializationSize()
         {
-            return sizeof(int);
+            return sizeof(short);
         }
         
     }
