+++ DofusProtocol/Types/Types/game/shortcut/ShortcutEmote.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 10/28/2014 16:38:06
+// Generated on 12/29/2014 21:14:45
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -23,7 +23,7 @@ namespace Stump.DofusProtocol.Types
         {
         }
         
-        public ShortcutEmote(int slot, byte emoteId)
+        public ShortcutEmote(sbyte slot, byte emoteId)
          : base(slot)
         {
             this.emoteId = emoteId;
