+++ Server/Stump.Server.WorldServer/Game/Actors/RolePlay/Characters/Character.cs
@@ -1964,12 +1964,12 @@ namespace Stump.Server.WorldServer.Game.Actors.RolePlay.Characters
 
         public void DisplayNotification(string text, NotificationEnum notification = NotificationEnum.INFORMATION)
         {
-            Client.Send(new NotificationByServerMessage((ushort) notification, new []{text}));
+            Client.Send(new NotificationByServerMessage((short) notification, new []{text}));
         }
 
         public void DisplayNotification(NotificationEnum notification, params object[] parameters)
         {
-            Client.Send(new NotificationByServerMessage((ushort)notification, parameters.Select(entry => entry.ToString())));
+            Client.Send(new NotificationByServerMessage((short)notification, parameters.Select(entry => entry.ToString())));
         }
 
         public void DisplayNotification(Notification notification)
@@ -2973,9 +2973,9 @@ namespace Stump.Server.WorldServer.Game.Actors.RolePlay.Characters
                 AlignmentValue,
                 AlignmentGrade,
                 CharacterPower,
-                Honor,
-                LowerBoundHonor,
-                UpperBoundHonor,
+                (short)Honor,
+                (short)LowerBoundHonor,
+                (short)UpperBoundHonor,
                 PvPEnabled ? (sbyte)1 : (sbyte)0);
         }
