+++ Server/Stump.Server.WorldServer/Game/Actors/RolePlay/Mounts/Mount.cs
@@ -431,7 +431,7 @@ namespace Stump.Server.WorldServer.Game.Actors.RolePlay.Mounts
         public MountInformationsForPaddock GetMountInformationsForPaddock()
         {
             var character = CharacterManager.Instance.GetCharacterById(OwnerId);
-            return new MountInformationsForPaddock(Model.Id, Name, character.Name);
+            return new MountInformationsForPaddock((sbyte)Model.Id, Name, character.Name);
         }
 
         #endregion
